🪲 [Fix]: Fix linter settings and docs (#2)
This pull request introduces minor improvements to documentation and
testing configuration. The most notable changes are an update to the
description in the `Get-PSModuleTest` function and the addition of
environment variable settings for the linter in the GitHub Actions
workflow.

Documentation improvements:

* Added a `.DESCRIPTION` section to the `Get-PSModuleTest` function for
better clarity and consistency in documentation.

Testing and CI configuration:

* Updated `.github/PSModule.yml` to include a `Linter` section with
environment variables that disable several validation checks,
streamlining the linter process in CI.